first step to the new API maybe someday we will know how it works, maybe not. The new paramter conference id may or may not be relevant The original URLs are removed which means media can't pull files anymore. New SCP logic is needed with the great potential to fuck media by overwriting anything. New 3 latter language may or may not work New logic for thumbnails is needed as they are also not pulled anymore
Started again (as there was no option to keep the dev and the 32c3 branch up with so much changes) to remove
all global variables in favor of a dict which represents the ticket.
This makes the script a litte more tracker dependet but other sources could still create such a dict.